requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Pillow==11.0.0
gunicorn==21.2.0
werkzeug==3.0.1